feat(task): adding user resource to requests and creating guard

+++ src/modules/task/task.service.ts
@@ -1,30 +1,50 @@
-import { Injectable, NotFoundException } from '@nestjs/common';
+import { Injectable, NotFoundException, UnauthorizedException } from '@nestjs/common';
 import { CreateTaskDto } from './dto/create-task.dto';
 import { UpdateTaskDto } from './dto/update-task.dto';
 import { Task } from './task.entity';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
 import { TaskStatus } from './dto/task.enum';
+import { User } from '../users/user.entity';
 
 @Injectable()
 export class TaskService {
   constructor(
     @InjectRepository(Task)
-    private taskRepository: Repository<Task>
+    private taskRepository: Repository<Task>,
+    @InjectRepository(User)
+    private userRepository: Repository<User>
   ){}
 
-  create(createTaskDto: CreateTaskDto): Promise<Task> {
-    return this.taskRepository.save(createTaskDto);
+  async create(createTaskDto: CreateTaskDto, user_email: string): Promise<Task> {
+    const user = await this.userRepository.findOne({
+      where: {
+        email: user_email
+      }
+    });
+
+    const { title, description, priority } = createTaskDto;
+    
+    const task = new Task();
+    task.title = title;
+    task.description = description;
+    task.priority = priority;
+    task.user = user;
+
+    return this.taskRepository.save(task);
   }
 
-  findAll(): Promise<Task[]> {
-    return this.taskRepository.find();
-  }
+  async findAll(user_email: string) {
+    const user = await this.userRepository.find({
+      where: {
+        email: user_email
+      }
+    });
 
-  async findAllByGroup() {
     const todo_list = await this.taskRepository.find({
       where: {
-        status: TaskStatus.TO_DO
+        status: TaskStatus.TO_DO,
+        user: user 
       },
       order: {
         position: 'ASC'
@@ -33,7 +53,8 @@ export class TaskService {
 
     const in_progress_list = await this.taskRepository.find({
       where: {
-        status: TaskStatus.IN_PROGRESS
+        status: TaskStatus.IN_PROGRESS,
+        user: user 
       },
       order: {
         position: 'ASC'
@@ -42,7 +63,8 @@ export class TaskService {
 
     const done_list = await this.taskRepository.find({
       where: {
-        status: TaskStatus.DONE
+        status: TaskStatus.DONE,
+        user: user
       },
       order: {
         position: 'ASC'
@@ -65,17 +87,20 @@ export class TaskService {
     ];
   }
 
-  async findOne(id: number): Promise<Task> {
-    const task = await this.taskRepository.findOneBy({id});
-    
-    if(!task){
-      throw new NotFoundException('Task not found');
+  async update(id: number, updateTaskDto: UpdateTaskDto, user_email: string): Promise<Task> {
+    const user = await this.userRepository.findOne({
+      where: {
+        email: user_email
+      },
+      relations: ['tasks']
+    });
+
+    const taskExists = user.tasks.some(task => task.id === id);
+
+    if (!taskExists) {
+      throw new UnauthorizedException('Task does not belong to the user');
     }
 
-    return task;
-  }
-
-  async update(id: number, updateTaskDto: UpdateTaskDto): Promise<Task> {
     const task = await this.taskRepository.preload({
       id,
       ...updateTaskDto,
@@ -88,7 +113,20 @@ export class TaskService {
     return this.taskRepository.save(task);
   }
 
-  async remove(id: number): Promise<void> {
+  async remove(id: number, user_email: string): Promise<void> {
+    const user = await this.userRepository.findOne({
+      where: {
+        email: user_email
+      },
+      relations: ['tasks']
+    });
+
+    const taskExists = user.tasks.some(task => task.id === id);
+
+    if (!taskExists) {
+      throw new UnauthorizedException('Task does not belong to the user');
+    }
+
     const result = await this.taskRepository.delete(id);
   
     if (result.affected === 0) {
